feat: TurboQuant Step 2 — Triton quantized decode attention kernel
Replace dequant-then-FlashAttention with a fused Triton decode kernel that
reads packed uint8 KV buffers directly. Supports both integer (3-bit) and
split-channel fractional (3.5-bit) bit-widths.

New components:
- Split-channel codec (quant_ops.py): 5 functions for fractional bit-widths
  via lo/hi channel groups with independent codebooks and rotation matrices
- Compact pool (turboquant_pool.py): KVCache subclass with packed uint8
  storage, ~4.3x compression vs BF16 (0.05 GB vs ~3.9 GB for Qwen3.5-9B)
- Triton decode kernel (turboquant_decode_attention.py): two-stage
  flash-decoding with inline bit extraction, codebook gather, QJL sign-bit
  correction for keys, and weighted V accumulation in rotated space
- TurboQuant attention backend (turboquant_backend.py): inherits
  TritonAttnBackend, overrides forward_decode for quantized path
- Auto-selection: --kv-cache-quantization turboquant auto-selects backend
- Tests: kernel correctness (7 tests) + server integration (6 tests)

Per-group QJL scale factor (sqrt(pi/2)/d_group) matches paper and MLX
reference implementation.

+++ python/sglang/srt/layers/quantization/turboquant/config.py
@@ -30,9 +30,9 @@ class TurboQuantConfig(QuantizationConfig):
     Enabled via --kv-cache-quantization turboquant.
     """
 
-    def __init__(self, bits: int = 3, seed: int = 42):
+    def __init__(self, bits: float = 3, seed: int = 42):
         super().__init__()
-        self.bits = int(bits)
+        self.bits = float(bits)
         self.seed = seed
 
     def get_name(self) -> str:
